Final version for a while
I'm halting all development on SIG as of now. This was only ever intended as a project to help me learn basic javascript. It's actually gone quite a bit further than I expected, but I still don't consider it a game. I don't think people would find this fun to play. I'm starting development on another game I've been wanting to make for years and will be doing it in javascript as well, if you want to see something I'm working on that I actually planned out in advance and didn't do it 100% by the seat of my pants, like SIG.